+++ android/src/newarch/CaptureProtectionModuleSpec.kt
